+++ client/package.json
@@ -30,7 +30,7 @@
     "build": "react-scripts build",
     "test": "react-scripts test",
     "eject": "react-scripts eject",
-    "lint": "eslint . --ext .ts",
+    "lint": "eslint . --ext .tsx",
     "prettier-format": "prettier --config ../.prettierrc 'src/**/*.ts' --write"
   },
   "eslintConfig": {
