+++ client/src/data/todos.ts
@@ -1,16 +1,16 @@
-export interface Todo {
-  id: number;
-  text: string;
-  isDone: boolean;
-  created: string;
-  expiringDate: string;
-}
+// export interface Todo {
+//   id: number;
+//   text: string;
+//   isDone: boolean;
+//   created: string;
+//   expiringDate: string;
+// }
 
-export interface TodoLists {
-  id: number;
-  listName: string;
-  items: Todo[];
-}
+// export interface TodoLists {
+//   id: number;
+//   listName: string;
+//   items: Todo[];
+// }
 
 export const todoLists = [
   {
