+++ package.json
@@ -1,41 +1,45 @@
 {
-	"name": "todo-list",
-	"version": "1.0.0",
-	"description": "",
-	"main": "index.js",
-	"scripts": {
-		"test": "echo \"Error: no test specified\" && exit 1",
-		"lint": "eslint . --fix --max-warnings=0",
-		"format": "prettier . --write",
-		"prepare": "husky install",
-		"lint-staged": {
-			"*.js": "npm run lint",
-			"*.{js,css,md,html,json}": "npm run format"
-		}
-	},
-	"repository": {
-		"type": "git",
-		"url": "git+https://github.com/alexlex7/todo-list.git"
-	},
-	"keywords": [],
-	"author": "",
-	"license": "ISC",
-	"bugs": {
-		"url": "https://github.com/alexlex7/todo-list/issues"
-	},
-	"homepage": "https://github.com/alexlex7/todo-list#readme",
-	"devDependencies": {
-		"eslint": "^8.31.0",
-		"eslint-config-airbnb": "^19.0.4",
-		"eslint-config-prettier": "^8.6.0",
-		"eslint-plugin-prettier": "^4.2.1",
-		"eslint-plugin-react": "^7.31.11",
-		"husky": "^8.0.3",
-		"lint-staged": "^13.1.0",
-		"prettier": "2.8.2"
-	},
-	"lint-staged": {
-		"*.js": "eslint --cache --fix",
-		"*.--write": "prettier --write"
-	}
+  "name": "todo-list",
+  "version": "1.0.0",
+  "description": "",
+  "main": "index.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "lint": "eslint . --fix --max-warnings=0",
+    "format": "prettier . --write",
+    "prepare": "husky install",
+    "lint-staged": {
+      "*.js": "npm run lint",
+      "*.{js,css,md,html,json}": "npm run format"
+    }
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/alexlex7/todo-list.git"
+  },
+  "keywords": [],
+  "author": "",
+  "license": "ISC",
+  "bugs": {
+    "url": "https://github.com/alexlex7/todo-list/issues"
+  },
+  "homepage": "https://github.com/alexlex7/todo-list#readme",
+  "devDependencies": {
+    "@typescript-eslint/eslint-plugin": "^5.48.1",
+    "@typescript-eslint/parser": "^5.48.1",
+    "eslint": "^8.31.0",
+    "eslint-config-airbnb": "^19.0.4",
+    "eslint-config-prettier": "^8.6.0",
+    "eslint-config-typescript": "^3.0.0",
+    "eslint-plugin-prettier": "^4.2.1",
+    "eslint-plugin-react": "^7.31.11",
+    "eslint-plugin-typescript": "^0.14.0",
+    "husky": "^8.0.3",
+    "lint-staged": "^13.1.0",
+    "prettier": "2.8.2"
+  },
+  "lint-staged": {
+    "*.ts": "eslint --cache --fix",
+    "*.--write": "prettier --write"
+  }
 }
