+++ server/.lintstagedrc.json
@@ -1,0 +1,1 @@
+{ "*.ts": "eslint --fix", "*.--write": "prettier --write" }
