+++ server/index.ts
@@ -1,0 +1,23 @@
+import express, { Express, Request, Response } from 'express';
+import dotenv from 'dotenv';
+import morgan from 'morgan';
+import path from 'path';
+
+// The dotenv package is used to read environment variables from the .env file.
+dotenv.config();
+
+const app: Express = express();
+const port = process.env.PORT;
+
+app.use(morgan('dev'));
+app.use(express.json());
+app.use(express.urlencoded({ extended: false }));
+app.use(express.static(path.join(__dirname, 'public')));
+
+app.get('/', (req: Request, res: Response) => {
+  res.send('Express + TypeScript Server is running');
+});
+
+app.listen(port, () => {
+  console.log(`⚡️[server]: Server is running at http://localhost:${port}`);
+});
