--- server/views/error.jade
@@ -1,6 +1,0 @@
-extends layout
-
-block content
-  h1= message
-  h2= error.status
-  pre #{error.stack}
